[#26137] VerilatorPlugin: Wishbone: Add option to extend waiting
The extended waiting is not made a default because it seems to break
some already developed Wishbone-based Verilated peripherals.

It definitely breaks the communication in Renode's liteuart sample,
which uses a LiteX-generated IP Core, so the issue might not be limited
to this single IP Core.

To use this option, `-DWISHBONE_EXTRA_WAIT_TICK` should be added to
the peripheral's `COMP_(LIB|EXEC)_ARGS` in its `CMakeLists.txt` file.