[#31486] Change extension of Arch package generated by fpm
Rationale:
As a result of the following commit:
https://github.com/jordansissel/fpm/commit/ca3477b67ba6bc9adc1cbe97e560061a739a12fc,
packages generated with `--pacman-compression xz` have an invalid extension: `.zst`
instead of `.xz`. Versions from 1.12.0 to 1.14.1 (at the moment of writing this)
are affected.

This commit is a workaround - if the user will generate packages with an
older fpm (<1.12.0), the file will remain unchanged. If an affected version of fpm will
be used - it will be renamed to have the expected file extension.