[#33538] conda: Fix accessibility on Windows
Renode.exe is installed to unusual Conda path on Windows:
`$CONDA_PREFIX/Library/renode/exec`, which isn't any of the default
environment ones added by Conda to PATH.

An activate script for Windows was installed properly but it turns out
`*.bat` are only used with the CMD shell on Windows. When used through
Bash, Conda calls `*_{de,}activate.sh` scripts. The new scripts modify
PATH during activation/deactivation of Conda environment.

On PowerShell, which is the most obvious way to use Conda on Windows,
`*_{de,}activate.ps1` scripts are used instead. However, the problem is
they can't be used to modify PATH.

Therefore, the `Renode.ps1` wrapper is installed to one of the Conda bin
paths to make running Conda-packaged Renode on PowerShell an easy task.