[GH#404] Us IF/ELSE in Robot Framework
Robot Framework added the IF/ELSE syntax in RF 4.0. Update places
the keywords Run Keyword If and Run Keyword Unless were used to
use this newer syntax. This allows the existing scripts to be
compatible with newer version of Robot Framework, such as 6.0
which marks the Run Keyword Unless keyword as deprecated and displays
a warning.