[#44214] gdb_compare: Add error detection for GDB responses
Not thoroughly tested - the currently defined patterns in RE_GDB_ERRORS
were chosen by connecting with gdb manually and trying as many illegal
things as possible to get some sample error responses.

In case of a match, an adequate message is printed alongside the
matching message, but execution continues. Raising an exception is not
worth it, because there is a risk of a false positive, which would mean
erroneously terminating the script's execution (which, historically,
could mean many hours of wasted time).

The last command sent to GDB is also remembered and printed, which is
useful in situations when GDB does not make the cause of the problem
obvious, e.g.:

	(gdb) step asdf
	No symbol "asdf" in the current context.

Also should also be useful for tweaking RE_GDB_ERRORS in case of false
positives.