[#45572] renode-test: Add -N/--retry flag
The -N flag controls 2 types of retrying strategies at once:
1) Test case retry - handled by the robotframework-retryfailed module
2) Suite retry - our own implementation

Suite retries are triggered only if a suite has failed AND conditions
are met to warrant a suite retry. Currently, those conditions are
finding a "\[Errno \d+\] Connection refused" regex match in a suite's
robot_output.xml messages, because it means that Robot's connection to
Renode was lost (most likely due to a crash) and Renode needs to be
restarted.

This means that the true upper bound for the number of times a test can
be started is the argument of -N squared (e.g. up to 25 times for -N5).