Harmonize LibraryVerilatorConnection with SocketVerilatorConnection
When setting the BaseVerilatedPeripheral.SimulationFilePath
Connect is always called after setting. This is required for sockets but
not for libraries. This causes a warning to be displayed whenever using
a library verilated peripheral. "The Verilated peripheral is already
connected." to mitigate this we can just require that connect is called
after setting the path (already done at least for
BaseVerilatedPeripheral) that way the warning only gets shown if Connect
is truly called twice.