[#63390] Vegaboard-RI5CY: Fix demo
The board seemed not to boot properly as nothing was displayed in
the analyzer and multiple warnings appeared in the logs.
It turned out that the UART had not been enabled properly as some
important registers were not mocked or were mocked incorrectly,
which resulted in Zephyr failing to enable and write to the UART.
I mocked the missing needed peripherals and got Zephyr to run correctly.