[#85591] BaseRiscV: Fix MTVEC/STVEC write and STVEC read adjustments
The behavior expected in the test had to be fixed because it depended on
Monitor adjustments which weren't in line with the ones made to software
MTVEC/STVEC accesses. Now the software and monitor cases for Auto Mode
are aligned (bit 1 should stay unset for priv v1.10 used in the test).