[#86311] arm: gic: Improvements to handling ARMv8.2 SMP platforms
This affects the way that GenericTimer is connected to CPUs.
For convenience, so far we ignored Affinity0 for ARMv8.2 cores.
So for ARMv8.2 CPU with `cpuId: 0x100` we could just do:
```
timer1: Timers.ARM_GenericTimer @ cpu1
    NonSecureEL2PhysicalTimerIRQ -> gic#1@26
```
Now it's no longer possible, but it's complemented by allowing hex values in timer connections.
So for the same scenario, we now have to do:
```
timer1: Timers.ARM_GenericTimer @ cpu1
    NonSecureEL2PhysicalTimerIRQ -> gic#0x100@26
```